OpenCloud->tryGetObject() local cache array
I noticed that the API request to get the DataObject was being
  called several times on the same basic logic set, on the same `$key`.

Each call resulted in a separate API request, with all the overhead & latency.

So I implemented a very simple local cache for the DataObject, on the Adapter.

The local cache of the DataObject is cleared on write or delete,
  otherwise it persists for a single execution (page load/script run) and no more;
  so it's not a presistant cache, just to prevent identical API requests on the same run.